Move plus minus button to its own file
This components will be reused in other parts of the app.
