+++ saplings/circuits/src/components/PlusMinusButton.scss
@@ -1,0 +1,33 @@
+/**
+ * Copyright 2018-2020 Cargill Incorporated
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+.minus-button, .plus-button {
+  padding: 0;
+  border: none;
+  outline: none;
+  height: fit-content;
+  cursor: pointer;
+  margin-left: 0.5rem;
+  background-color: transparent;
+}
+
+.plus-button {
+  color: var(--color-primary-light);
+}
+
+.minus-button {
+  color: var(--color-attention);
+}
